Create a presentational component App
